% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conord.R
\name{conord}
\alias{conord}
\title{Plots quality assessment}
\usage{
conord(PS, formula, method = c("CCA", "RDA"), facets, scaling = 2,
  tax_level = "Phylum", tax_n = 7)
}
\arguments{
\item{PS}{(required) asdf}

\item{formula}{(required) asdf}

\item{method}{asdf. Defaults to CCA.}

\item{facets}{asdf}

\item{scaling}{asdf. Defaults to 2.}

\item{tax_level}{asdf. Defaults to Phylum.}

\item{tax_n}{asdf. Defaults to 7.}
}
\value{
A ggplot object.
}
\description{
Description paragraph
}
\examples{
\dontrun{
conord(PS,~ SL_NPOC + SL_NO3 + SL_NH4,facets=~reactor.cat,tax_level='Class',scaling=2,method='RDA')
}

}
\references{
X
Y
Z
}
\seealso{
\code{\link[phyloseq]{ordinate}} \code{\link[vegan]{cca}} \code{\link[vegan]{rda}}
}
