% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constord.R
\name{constord}
\alias{constord}
\title{Plots constrained ordination results}
\usage{
constord(PS, formula, method = c("CCA", "RDA"), facets, scaling = 2,
  tax_level = "Phylum", tax_n = 7)
}
\arguments{
\item{PS}{(required) A phyloseq object.}

\item{formula}{(required) Right-hand side of the model formula starting with a tilde ("~").}

\item{method}{Constrained ordination method to be applied. User may choose Constrained Correspondence Analysis (CCA) or Redundancy Analysis (RDA). Defaults to CCA.}

\item{facets}{Variable in sample_data(PS) to facet the plot by. Statement starts with a tilde ("~").}

\item{scaling}{Scaling for species and site/sample scores in biplot. Options are the same as those found in the \code{\link[vegan]{scores}} function: "species" scaling (1) or "site" scaling (2). The user should designate the appropriate scaling for thier intended analysis. Further information regarding scaling can be found in the Details below. Defaults to 2.}

\item{tax_level}{Taxonomic level to represent species composition using color. Defaults to "Phylum".}

\item{tax_n}{The number of taxonomic groups to identify using color (at the taxonomic level 'tax_level'). The most abundant tax_n will be selected. All other taxonomic groups will be collapsed into an additional "Other" category for visualization. Defaults to 7.}
}
\value{
A ggplot object.
}
\description{
Function conord performs Constrained Correspondence Analysis (CCA) or Redundancy Analysis (RDA) and plots the results as a ggplot object.
}
\details{
\subsection{Scaling}{
  Species scaling (1) results in a distance biplot. The distance biplot is intended to enable the user to interpret the relationships between sites/samples.
  Site scaling (2) results in correlation biplot. The correlation biplot enables the user to interpret the correlation between descriptors (species) within the ordination. Positions of sites/samples are not approximations of thier true locations; use species scaling (1) to interpret site/samples.
  A much more complete discussion of the implications of scaling (and intepretation of the ordination results) is provided in Legendre and Legendre (1998, p. 403-404, 585-587).
}
}
\examples{
\dontrun{
constord(PS,~ SL_NPOC + SL_NO3 + SL_NH4,facets=~reactor.cat,tax_level='Class',scaling=2,method='RDA')
}

}
\references{
Legendre, P. and Legendre, L. (1998) Numerical Ecology. 2nd English ed. Elsevier.
}
\seealso{
\code{\link[phyloseq]{ordinate}} \code{\link[vegan]{cca}} \code{\link[vegan]{rda}} \code{\link[vegan]{scores}}
}
