% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.R
\name{cont}
\alias{cont}
\title{Plots quality assessment}
\usage{
cont(x, y, z, data, method = c("linear", "spline", "loess"), ...)
}
\arguments{
\item{x}{(required) asdf}

\item{y}{(required) asdf}

\item{z}{(required) asdf}

\item{data}{(required) asdf}

\item{method}{asdf. Defaults to linear.}

\item{...}{Additional arguments for loess.}
}
\value{
A ggplot object.
}
\description{
Description paragraph
}
\examples{
\dontrun{
cont('xvar','yvar','zvar',dat,method='loess',degree = 2, span = 0.1)
}

}
\references{
X
Y
Z
}
\seealso{
\code{\link[stats]{loess}} \code{\link[splancs]{inpip}} \code{\link[akima]{interp}}
}
