% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contplot.R
\name{contplot}
\alias{contplot}
\title{Plots a color/contour plot of a three dimensional surface}
\usage{
contplot(x, y, z, data, method = c("linear", "spline", "loess"),
  removeMissing = FALSE, ...)
}
\arguments{
\item{x}{(required) Vector of x-coordinates (within 'data').}

\item{y}{(required) Vector of y-coordinates (within 'data').}

\item{z}{(required) Vector of z-coordinates (within 'data').}

\item{data}{(required) A dataframe-like object containing x, y, and z.}

\item{method}{Surface fitting method. Options are 'linear', 'spline', and
'loess' (see Details). Defaults to 'linear'.}

\item{removeMissing}{Remove entries with missing values (see Details).
Defaults to FALSE.}

\item{...}{Additional arguments for loess.}
}
\value{
A ggplot object.
}
\description{
Function contplot fits a surface through a 3-dimensional irregularly-spaced
dataset and presents the fitted surface as a 2-dimensional color/contour
plot. 'linear', 'spline', and 'loess' fitting methods are currently supported
for this function. Please note that missing values in 'data' are not
tolerated (see Details).
}
\details{
This function can be used for both regularly and irregularly spaced data.
\subsection{'method' - surface fitting methods}{
\itemize{
  \item 'linear' uses the \code{\link[akima]{interp}} function in the akima R
    package (Akima and Gebhardt, 2016) to carry out bivariate linear
    interpolation.
  \item 'spline' also uses the \code{\link[akima]{interp}} function in the
    akima R package (Akima and Gebhardt, 2016), but the 'linear' argument
    passed to \code{\link[akima]{interp}} is set to 'FALSE', resulting in
    bicubic spline interpolation.
  \item 'loess' uses the \code{\link[stats]{loess}} (R Core Team, 2017) and
    \code{\link[splancs]{inpip}} (Rowlingson and Diggle, 2017) functions to
    carry out locally weighted scatterplot smoothing (LOESS).
}
}

\subsection{Missing Values}{
Missing values in 'x', 'y', or 'z' are not tolerated by these functions and
  will result in an error. Setting 'removeMissing' to TRUE will remove any
  and all entries that contain missing values before interpolation is
  carried out.
}
}
\examples{
\dontrun{
library(theseus)
data(akima, package='akima')
str(akima)
dat <- data.frame(cbind(a=akima$x, b=akima$y, c=akima$z))
p.cp.l <- contplot(x='a', y='b', z='c', data=dat, method='linear')
p.cp.l + ggtitle("contour plot with linear surface fitting")
p.cp.s <- contplot(x='a', y='b', z='c', data=dat, method='spline')
p.cp.s + ggtitle("contour plot with spline surface fitting")
p.cp.loess <- contplot(x='a', y='b', z='c', data=dat, method='loess')
p.cp.loess + ggtitle("contour plot with loess surface fitting")
}

}
\references{
Akima, H., and Gebhardt, A. (2016). akima: Interpolation of Irregularly and
  Regularly Spaced Data. R package version 0.6-2.
  \url{https://CRAN.R-project.org/package=akima}.
R Core Team (2017). R: A language and environment for statistical computing.
  R Foundation for Statistical Computing, Vienna, Austria.
  \url{https://www.R-project.org/}.
Rowlingson, B., and Diggle, P. (2017). splancs: Spatial and Space-Time
  Point Pattern Analysis. R package version 2.01-40.
  \url{https://CRAN.R-project.org/package=splancs}.
}
\seealso{
\code{\link[stats]{loess}} \code{\link[splancs]{inpip}}
  \code{\link[akima]{interp}}
}
