% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstoveg_otu.R
\name{pstoveg_otu}
\alias{pstoveg_otu}
\title{converts the otu_table slot of a phyloseq object to a vegan-compatible matrix}
\usage{
pstoveg_otu(PS)
}
\arguments{
\item{PS}{(required) a phyloseq object}
}
\value{
A matrix containing a phyloseq object's otu_table slot.
}
\description{
physeq2veg_otu is a helper function intended to convert the species/taxa count slot of a phyloseq object to a vegan-friendly matrix.
}
\examples{
\dontrun{
library(phyloseq)
data("GlobalPatterns")
# inspect otu_table()
dim(otu_table(GlobalPatterns))
str(otu_table(GlobalPatterns))
taxa_are_rows(GlobalPatterns)
gp.otu <- physeq2veg_otu(GlobalPatterns)
dim(gp.otu)
str(gp.otu)
}

}
\seealso{
\code{\link[phyloseq]{phyloseq-class}} \code{\link[phyloseq]{otu_table-class}} \code{\link[phyloseq]{otu_table}}
}
