% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual.R
\name{qual}
\alias{qual}
\title{Plots quality assessment}
\usage{
qual(f_path, r_path, idx, percentile = 0.25, n_trim = 100, n_seqs = 12,
  q = c(25, 30, 35), bins = 50, nc = 1,
  seed = sample.int(.Machine$integer.max, 1), verbose = FALSE)
}
\arguments{
\item{f_path}{(required) asdf}

\item{r_path}{(required) asdf}

\item{idx}{asdf}

\item{percentile}{asdf. Defaults to .25.}

\item{n_trim}{asdf. Defaults to 100.}

\item{n_seqs}{asdf. Defaults to 12.}

\item{q}{asdf. Defaults to 25, 30, and 35.}

\item{bins}{adsf. Defaults to 50.}

\item{nc}{asdf. Defaults to 1.}

\item{seed}{asdf.}

\item{verbose}{asdf. Defaults to FALSE.}
}
\value{
A ggplot object with the following attributes:
\describe{
\item{idx}{asdf}
\item{seed}{asdf}
}
}
\description{
Description paragraph
}
\examples{
\dontrun{
fns <- sort(list.files(data_path,full.names=TRUE))
f_path <- fns[grepl('R1', fns)]
r_path <- fns[grepl('R2', fns)]
p1 <- qual(f_path,r_path,n_seqs=12,verbose=TRUE,percentile=.25,nc=12)
p1 + geom_hline(yintercept=175) + geom_vline(xintercept=275)
}

}
\references{
X
Y
Z
}
\seealso{
\code{\link[ShortRead]{qa}}
}
