% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualcontour.R
\name{qualcontour}
\alias{qualcontour}
\title{Plots quality assessment}
\usage{
qualcontour(f_path, r_path, idx, percentile = 0.25, n_trim = 100,
  n_samples = 12, q = c(25, 30, 35), bins = 50, nc = 1,
  seed = sample.int(.Machine$integer.max, 1), verbose = FALSE)
}
\arguments{
\item{f_path}{(required) A character vector locating the forward read (Read 1) .fastq files}

\item{r_path}{(required) A character vector locating the reverse read (Read 2) .fastq files}

\item{idx}{Indexes (within f_path and r_path) identifying specific .fastq files for analysis}

\item{percentile}{The percentile to be targeted . Defaults to .25 (i.e. the first quartile).}

\item{n_trim}{asdf. Defaults to 100.}

\item{n_samples}{Integer indicating the number of samples to include in the visualization. Defaults to 12.}

\item{q}{A numeric vector designating Phred quality scores to be represented on the plot. Defaults to 25, 30, and 35.}

\item{bins}{Integer designating the number of bins each read should be seperated into. For example, visualizing a 250 bp read with 50 bins would imply that each bin represents 5 cycles/bp. Increasing the number of bins improves granularity at the cost of memory and processing speed. Defaults to 50.}

\item{nc}{The number of cores to use when multithreading. Defaults to 1.}

\item{seed}{An integer value to be used when randomly selecting the subset of samples to be visualized.}

\item{verbose}{If set to TRUE, provides verbose output. Defaults to FALSE.}
}
\value{
A ggplot object with the following attributes:
\describe{
\item{idx}{Samples used to generate the plot.}
\item{seed}{Seed used to select the samples used to generate the plot.}
}
}
\description{
This function generates a 2-D contour map representing the location of quality scores for a designated percentile. It is intended to assist with deciding where trimming should be performed.
}
\examples{
\dontrun{
fns <- sort(list.files(data_path,full.names=TRUE))
f_path <- fns[grepl('R1', fns)]
r_path <- fns[grepl('R2', fns)]
p1 <- qualcontour(f_path,r_path,n_samples=12,verbose=TRUE,percentile=.25,nc=1)
p1 + geom_hline(yintercept=175) + geom_vline(xintercept=275)
}

}
\references{
X
Y
Z
}
\seealso{
\code{\link[ShortRead]{qa}} \code{\link[dada2]{plotQualityProfile}}
}
