% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{tax_table<-}
\alias{reexports}
\alias{otu_table<-}
\alias{reexports}
\alias{sample_data<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{phyloseq}{\code{\link[phyloseq]{tax_table<-}}, \code{\link[phyloseq]{otu_table<-}}, \code{\link[phyloseq]{sample_data<-}}}
}}

