% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_seqtab.2vs3.R
\docType{data}
\name{sigtab.2vs3}
\alias{sigtab.2vs3}
\title{Differential abundance data}
\format{A 205x13 dataframe containing DESeq2 differential abundance data.}
\usage{
sigtab.2vs3
}
\description{
Differential abundance data
}
\keyword{datasets}
