% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_seqtab.R
\docType{data}
\name{sigtab}
\alias{sigtab}
\title{Differential abundance data}
\format{A 137x13 dataframe containing DESeq2 differential abundance data.}
\usage{
sigtab
}
\description{
Differential abundance data
}
\keyword{datasets}
