% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf.R
\name{surf}
\alias{surf}
\title{Plots quality assessment}
\usage{
surf(PS, covariates)
}
\arguments{
\item{PS}{(required) asdf}

\item{covariates}{(required) asdf}
}
\value{
A ggplot object.
}
\description{
Description paragraph
}
\examples{
\dontrun{
covariates <- c('SL_NPOC','SL_NO3','SL_NH4')
surf(PS,covariates)
}

}
\references{
X
Y
Z
}
\seealso{
\code{\link[vegan]{rda}} \code{\link[vegan]{envfit}} \code{\link[vegan]{ordisurf}}
}
