% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_WWTP_Impact.R
\docType{data}
\name{WWTP_Impact}
\alias{WWTP_Impact}
\title{Freshwater stream microbiome data}
\format{A phyloseq object containing sample_data, otu_table, tax_table, and phy_tree slots.}
\source{
\url{https://github.com/JacobRPrice/WWTP_Impact_on_Stream}
}
\usage{
WWTP_Impact
}
\description{
This dataset contains the results from twelve (12) 16S rDNA amplicon sequencing measurements collected from Wissahickon Creek and Sandy Run in southeastern Pennsylvania. Sequencing data was processed as described in Price et. al. (2016), and combined with chemical data to create a phyloseq object.
}
\references{
Price, J. R., Ledford, S. H., Ryan, M. O., Toran, L., Sales, C. M. (2017). Wastewater treatment plant effluent introduces recoverable shifts in microbial community composition in receiving streams (In Revision, Submitted June 27, 2017)
}
\keyword{datasets}
