% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contplot.R
\name{cont}
\alias{cont}
\title{Plots a contour plot of a three dimensional surface}
\usage{
cont(x, y, z, data, method = c("linear", "spline", "loess"),
  removeMissing = FALSE, ...)
}
\arguments{
\item{x}{(required) Vector of x-coordinates (within 'data').}

\item{y}{(required) Vector of y-coordinates (within 'data').}

\item{z}{(required) Vector of z-coordinates (within 'data').}

\item{data}{(required) A dataframe-like object containing x, y, and z.}

\item{method}{Surface fitting method. Options are "linear", "spline", and "loess". Defaults to linear.}

\item{removeMissing}{Remove entries with missing values. Defaults to FALSE.}

\item{...}{Additional arguments for loess.}
}
\value{
A ggplot object.
}
\description{
Function cont represents a 3 dimensional surface as a contour plot.
}
\examples{
\dontrun{
cont('xvar','yvar','zvar',dat,method='loess',degree = 2, span = 0.1)
}

}
\references{
X
Y
Z
}
\seealso{
\code{\link[stats]{loess}} \code{\link[splancs]{inpip}} \code{\link[akima]{interp}}
}
